scripts/aws_cleanup/requirements-dev.txt
boto3-stubs
boto3-stubs[ec2]